move is_executable to nose.util
